Revolutionary chinese propaganda, the colour of blue...
Added some notes to list-test.c, fixed the test makefile to stop it from
rebuilding the emulator lib ALLLL the time. Added a few methods to list-test for
appending to a list: append, and grabbing something off the end of the list: tail-pull.

These seem to work ok on a simple test. Valgrind -> no leaks. This container needs more work if it is to act as a tree however, each node would need an is-leaf flag and various othere pointers