changed covariance_fn quite a bit internally although the calls should remain in form at least.
The new covariance function is

Covariance function C(x1,x2,Theta)
	= theta1*Exp(Sum( (x1(i)-x2(i))^alpha ) * theta3) +
delta(x1,x2)*theta2

The large scale structure is still gaussian-esque although now
we can change the power of the exponent by #defiing Alpha in the
file emulator.c, theta3 is also raised to alpha in the calculation to
preserve dimensions.

However the old theta2 offset parameter is gone, as it was declared to
be somewhat suspect.

NOte that as of this commit the gradient descent algorithm has not been
updated to take into account the changed covariance function and
concomittant changed derivatives (although this should be easy in this
case)