work on estimate_threaded:
inserted a fn to produce a hex rep of pthread_self() for debugging.

maxlbfgs was checking -1*evalFnLBFGS(...) against the stored best value
while the fns in estimate threaded were checking evalFnLBFGS(...) which
resulted in configurations which were not optimal winning.

inserted (double)my_best into thread params struct, this is kept as a local running
check to make sure that the global "best" is one of these values.