multimin now uses a fixed sigma
with a fixed sigma multimin now optimises the length scales and the
noise parameter. this seems to work pretty well. It should be tested but
it appears that the final variance is a lot larger than before (no more
overfitting) but that the mean usually sticks nicely to the function.

so it looks like we've gotten a much more reasonable estimator