removed dep on global fnptrs in estimation
global fnptrs didnt work well on os-x, something was clobbering their
values, removed dep for most of the estimation code by using the fnptrs
defined in xmodel, included in the estimate_thetas_params structs.

need to test the R lib a bit more, but it should be possible to remove a
lot of cruft and pull out the dep on the global fnptrs entirely quite
soon.